Sample interaction framework start
This is the first half of a framework for testing interactions
for expected output - a tool to grab traces from kubectl logs,
a place to put them, and thats about it.

I consider all this unstable, so would like to evolve it in
subsequent patches (allowing others to collaborate) rather than
polishing this little step.